Cache reflected type information to speed up serialization  (#33)
Address performance hit caused by heavy use of reflection during serialization step using a memoization cache. The change improves serialization performance significantly (approximately a 4x factor) on a synthetic benchmark example, mileage may vary:

    - New Solution [Batch Size: 500000, Time in Milliseconds: 71586, 15.9MB Usage during test]
    - Current FSharpLu.Json [Batch Size: 500000, Time in Milliseconds: 208828, 15.9MB Usage during test]
    - Using delegate for option [Batch Size: 500000, Time in Milliseconds: 93371-94000] with delegate for option (Didn't observe RAM of process)

This yields a 2.9x speedup approx on the benchmark above. Based on the above, the `delegate` implementation was not retained. No significant RAM increase was observed in all three tests (just checked Task Manager unscientifically). The space requirement should go up linearly with the number of types to serialize in your application but with just the few types in the benchmark that was not noticeable.

Example type used for the benchmark:
```
type TestUnion = 
    | OptionOne of int
    | OptionTwo of string

type TestType = {
    FieldOne: string
    FieldTwo: int
    FieldThree: int option
    FieldFour: TestUnion
    FieldFive: string option option
    FieldSix: TestUnion option
}
```