Clearing TokenCache when expiring tokens are found
By default, `AuthenticationContext.AcquireTokenAsync()` generates tokens and stores them in an in-memory TokenCache, and then retrieves them from that cache when a token for the same principal/context is requested. Usually, tokens are generated with a RefreshToken that allows this function to check the expiration and regenerate the token when it's expired. In our case, we're using an override of .AcquireTokenAsync that doesn't create the RefreshToken, so the function doesn't automatically regenerate the token, and that's the cause of the BVT failures.
This change adds code to check the expiration time of tokens, and if one is found to expire within 10 minutes, the TokenCache is cleared, and a new token is generated.